+++ src/device/device.mstest/Tests/Asserts_Visa.cs
@@ -2,6 +2,38 @@ namespace cc.isr.VI.Device.Tests;
 
 public sealed partial class Asserts
 {
+    #region " visa implementation checks "
+
+    /// <summary>   Assert visa implementation should be loaded. </summary>
+    /// <remarks>   2025-09-18. </remarks>
+    /// <param name="verbose">  (Optional) True to verbose. </param>
+    public static void AssertVisaImplementationShouldBeLoaded( bool verbose = true )
+    {
+        if ( typeof( Ivi.Visa.IMessageBasedSession ).Assembly is System.Reflection.Assembly iviVisaAssembly )
+        {
+            if ( verbose )
+            {
+                Console.WriteLine( $"\t{iviVisaAssembly.FullName}." );
+                Console.WriteLine( $"\t{iviVisaAssembly.Location}." );
+            }
+        }
+        else
+            Assert.Fail( $"{nameof( Ivi.Visa.IMessageBasedSession )} assembly not found." );
+
+        if ( typeof( cc.isr.Visa.Gac.Vendor ).Assembly is System.Reflection.Assembly vendorVisaAssembly )
+        {
+            if ( verbose )
+            {
+                Console.WriteLine( $"\t{vendorVisaAssembly.FullName}." );
+                Console.WriteLine( $"\t{vendorVisaAssembly.Location}." );
+            }
+        }
+        else
+            Assert.Fail( $"{nameof( cc.isr.Visa.Gac.Vendor )} VISA assembly not found." );
+    }
+
+    #endregion
+
     #region " visa resource checks "
 
     /// <summary>   Assert visa resource manager should include a resource. </summary>