+++ src/vi/k2600/k2600.mstest/Contacts/ContactCheckTests.cs
@@ -41,6 +41,7 @@ public class ContactCheckTests : Device.Tests.Base.ContactCheckTests
     public override void InitializeBeforeEachTest()
     {
         Console.WriteLine( $"{this.TestContext?.FullyQualifiedTestClassName}: {DateTime.Now} {System.TimeZoneInfo.Local}" );
+        cc.isr.VI.Device.Tests.Asserts.AssertVisaImplementationShouldBeLoaded();
         Console.WriteLine( $"\tTesting {typeof( MeasureResistanceSubsystem ).Assembly.FullName}" );
 
         // create an instance of the session logger.