+++ src/vi/k2600/k2600.mstest/Source/CurrentSourceTests.cs
@@ -39,6 +39,7 @@ public class CurrentSourceTests : Device.Tsp.Tests.Base.TestBase
     public override void InitializeBeforeEachTest()
     {
         Console.WriteLine( $"{this.TestContext?.FullyQualifiedTestClassName}: {DateTime.Now} {System.TimeZoneInfo.Local}" );
+        cc.isr.VI.Device.Tests.Asserts.AssertVisaImplementationShouldBeLoaded();
         Console.WriteLine( $"\tTesting {typeof( cc.isr.VI.Tsp.K2600.CurrentSourceSubsystem ).Assembly.FullName}" );
 
         // create an instance of the session logger.