+++ src/vi/k2600/k2600.mstest/Subsystems/SubsystemsTests.cs
@@ -41,6 +41,7 @@ public class SubsystemsTests : Device.Tests.Base.SubsystemsTests
     public override void InitializeBeforeEachTest()
     {
         Console.WriteLine( $"{this.TestContext?.FullyQualifiedTestClassName}: {DateTime.Now} {System.TimeZoneInfo.Local}" );
+        cc.isr.VI.Device.Tests.Asserts.AssertVisaImplementationShouldBeLoaded();
         Console.WriteLine( $"\tTesting {typeof( cc.isr.VI.Tsp.K2600.MeasureResistanceSubsystem ).Assembly.FullName}" );
 
         // create an instance of the session logger.