+++ src/vi/k2600/k2600.mstest/Visa/ParseTests.cs
@@ -39,6 +39,7 @@ public class ParseTests : Device.Tests.Base.ParseTests
     public override void InitializeBeforeEachTest()
     {
         Console.WriteLine( $"{this.TestContext?.FullyQualifiedTestClassName}: {DateTime.Now} {System.TimeZoneInfo.Local}" );
+        cc.isr.VI.Device.Tests.Asserts.AssertVisaImplementationShouldBeLoaded();
         Console.WriteLine( $"\tTesting {typeof( cc.isr.VI.Tsp.K2600.K2600Device ).Assembly.FullName}" );
 
         // create an instance of the session logger.